app/requirements.txt
streamlit==1.45.1
redis==3.2.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
sentence-transformers==2.2.2
google-generativeai==0.8.5
torch==2.0.1
transformers==4.28.1
huggingface_hub==0.16.4
numpy==1.23.1